// ui/src/lib/api/generated/model/lookDetail.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Centoire API
 * AI-powered fashion platform API
 * OpenAPI spec version: 1.0.0
 */
import type { Garment } from './garment';

export interface LookDetail {
  look_id: string;
  /** @nullable */
  brand: string | null;
  /** @nullable */
  collection_id: string | null;
  /** @nullable */
  runway_url: string | null;
  /** @nullable */
  silhouette: string | null;
  /** @nullable */
  whole_look_flat: string | null;
  garments: Garment[];
  /** @nullable */
  status: string | null;
}